+++ frontend_tests/node_tests/recent_topics.js
@@ -103,6 +103,7 @@ mock_esm("../../static/js/muted_topics", {
 const narrow = mock_esm("../../static/js/narrow", {
     set_narrow_title: noop,
     hide_mark_as_read_turned_off_banner: noop,
+    handle_middle_pane_transition: noop,
     has_shown_message_list_view: true,
 });
 mock_esm("../../static/js/recent_senders", {
@@ -265,7 +266,7 @@ function get_topic_key(stream_id, topic) {
 
 function generate_topic_data(topic_info_array) {
     // Since most of the fields are common, this function helps generate fixtures
-    // with non common fields.
+    // with non-common fields.
     $.clear_all_elements();
     const data = [];