+++ static/js/compose.js
@@ -18,6 +18,7 @@ import {$t, $t_html} from "./i18n";
 import * as loading from "./loading";
 import * as markdown from "./markdown";
 import * as message_edit from "./message_edit";
+import * as narrow from "./narrow";
 import * as notifications from "./notifications";
 import {page_params} from "./page_params";
 import * as people from "./people";
@@ -100,6 +101,11 @@ export function update_fade() {
     compose_validate.warn_if_topic_resolved();
     compose_fade.set_focused_recipient(msg_type);
     compose_fade.update_all();
+}
+
+function update_on_recipient_change() {
+    update_fade();
+    compose_actions.update_narrow_to_recipient_visibility();
 }
 
 export function abort_xhr() {
@@ -394,10 +400,10 @@ export function initialize() {
     $("#below-compose-content .video_link").toggle(compute_show_video_chat_button());
     $(
         "#stream_message_recipient_stream,#stream_message_recipient_topic,#private_message_recipient",
-    ).on("keyup", update_fade);
+    ).on("keyup", update_on_recipient_change);
     $(
         "#stream_message_recipient_stream,#stream_message_recipient_topic,#private_message_recipient",
-    ).on("change", update_fade);
+    ).on("change", update_on_recipient_change);
     $(
         "#stream_message_recipient_topic",
     ).on("keyup", (event) => {
@@ -708,6 +714,11 @@ export function initialize() {
         e.stopPropagation();
 
         compose_ui.make_compose_box_full_size();
+    });
+
+    $("#compose").on("click", ".narrow_to_compose_recipients", (e) => {
+        e.preventDefault();
+        narrow.to_compose_target();
     });
 
     $("#compose").on("click", ".collapse_composebox_button", (e) => {