+++ zproject/computed_settings.py
@@ -232,7 +232,6 @@ if not TORNADO_PORTS:
 TORNADO_PROCESSES = len(TORNADO_PORTS)
 
 RUNNING_INSIDE_TORNADO = False
-AUTORELOAD = DEBUG
 
 SILENCED_SYSTEM_CHECKS = [
     # auth.W004 checks that the UserProfile field named by USERNAME_FIELD has
@@ -341,7 +340,7 @@ MEMCACHED_PASSWORD = get_secret("memcached_password")
 
 CACHES = {
     "default": {
-        "BACKEND": "django_bmemcached.memcached.BMemcached",
+        "BACKEND": "zerver.lib.singleton_bmemcached.SingletonBMemcached",
         "LOCATION": MEMCACHED_LOCATION,
         "OPTIONS": {
             "socket_timeout": 3600,